+++ src/main/controllers/TimetableController.ts
@@ -2,54 +2,70 @@ import { Controller } from "@peregrine/webserver"
 import { HttpErrors } from "@peregrine/exceptions"
 import Ride from "../models/Ride"
 import Repository from "../datasource/Repository";
+import Operator from "../models/Operator";
+import { Entity } from "../datasource/mongo/Entity";
 
 type json = {[key: string]: any}
 
-export default class TimetableController implements Controller<Ride> {
+export default class TimetableController implements Controller<Ride, Entity<Operator>> {
     public resourceName: string = "rides"
 
     constructor(protected repo: Repository<Ride>){}
     
-    public async get(id: string, _params: json): Promise<Ride> {
+    public async get(id: string, _params: json, _auth?: Entity<Operator>): Promise<Ride> {
         const model = await this.repo.getById(id)
-        if(model) return model
-        else throw new HttpErrors.Client.NotFound()
+        if(!model) throw new HttpErrors.Client.NotFound()
+        return model
     }
 
-    public async getAll(params: json): Promise<Ride[]> {
+    public async getAll(params: json, _auth?: Entity<Operator>): Promise<Ride[]> {
         const list = await this.repo.getAll()
         return params.stopId ? list.filter(item => item.stops.some(stop => stop.stop == params.stopId)) : list
     }
 
-    public async create(model: json, _params: json): Promise<Ride> {
-        TimetableController.validateModel(model)
-        return await this.repo.create(model as Ride)
+    public async create(model: json, _params: json, auth?: Entity<Operator>): Promise<Ride> {
+        if(!auth) throw new HttpErrors.Client.Unauthorised()
+        const ride = this.validateModel(model)
+        ride.operator = auth._id
+        return await this.repo.create(ride)
     }
 
-    public async update(id: string, model: json, _params: json): Promise<void> {
-        TimetableController.validateModel(model)
-        await this.repo.update(id, model as Ride)
+    public async update(id: string, model: json, _params: json, auth?: Entity<Operator>): Promise<void> {
+        if(!auth) throw new HttpErrors.Client.Unauthorised()
+        if(!this.hasEditRights(id, auth)) throw new HttpErrors.Client.Forbidden()
+        const ride = this.validateModel(model)
+        ride.operator = auth._id
+        await this.repo.update(id, ride)
     }
 
-    public updateAll(_model: json, _params: json): void | Promise<void> {
+    public updateAll(_model: json, _params: json, _auth?: Entity<Operator>): void | Promise<void> {
         throw new HttpErrors.Client.MethodNotAllowed()
     }
 
-    public async delete(id: string, _params: json): Promise<void> {
+    public async delete(id: string, _params: json, auth?: Entity<Operator>): Promise<void> {
+        if(!auth) throw new HttpErrors.Client.Unauthorised()
+        if(!this.hasEditRights(id, auth)) throw new HttpErrors.Client.Forbidden()
         await this.repo.delete(id)
     }
 
-    public deleteAll(_params: json): void | Promise<void> {
+    public deleteAll(_params: json, _auth?: Entity<Operator>): void | Promise<void> {
         throw new HttpErrors.Client.MethodNotAllowed()
     }
 
-    protected static validateModel(model: json){
-        if( (!model.operator || typeof model.operator != "string") || 
+    protected async hasEditRights(id: string, auth: Entity<Operator>): Promise<boolean> {
+        const model = await this.repo.getById(id)
+        return !!model && model.operator == auth._id
+    }
+
+    protected validateModel(model: json): Ride {
+        if( (model.operator) || 
             (!model.type || typeof model.type != "string") || 
             ( model.line && typeof model.line != "number") || 
             (!model.stops) || 
             (!model.excludeDays) || 
             (!model.departures) )
             throw new HttpErrors.Client.BadRequest()
+        
+        return model as Ride
     }
 }
